requirements.txt
discord.py
pytest
python-dotenv
celery
pika
flake8
